Changes for responsive
