+++ src/App.css
@@ -1,18 +1,14 @@
-*,*::after,*::before {
-    box-sizing: border-box;
-}
-
-body{
+#main{
     width: 100vw;
     height: 100vh;
-    margin: auto;
-    background: url('./assets/images/background-pattern-desktop.svg');
+    background: url('../src/assets/images/background-pattern-desktop.svg');
     background-position: top;
-    background-size: 100%;
     background-repeat: no-repeat;
-    background-color: hsl(275, 100%, 97%);
+    background-size: 100%;
 }
 
 .grid {
     display: flex;
+    justify-content: center;
+    align-items: center;
 }
