--- src/components/Container.css
@@ -1,44 +1,0 @@
-.container {
-    width: 900px;
-    height: auto;
-    margin: 150px auto 0 auto;
-    align-items: center;
-    justify-content: center;
-    border-radius: 20px;
-    background-color: hsl(0, 0%, 100%);
-    box-shadow: rgba(100, 100, 111, 0.2) 0px 7px 29px 0px;
-    padding: 15px 50px;
-}
-
-header {
-    font-size: 7rem;
-    display: inline-flex;
-    width: 100%;
-    align-items: center;
-}
-
-header img {
-    height: 4rem;
-    width: 2.5rem;
-    margin-right: 2.5rem;
-}
-
-header h1 {
-    font-size: 4rem;
-}
-
-.faq-list{
-    width: 100%;
-    height: 25%;
-    list-style: none;
-    display: grid;
-    grid-gap: 10px;
-    padding-left: 0px;
-    margin-bottom: 2rem;
-}
-
-.faq-list article:not(:last-child) {
-    border-bottom: 2px solid hsl(275, 100%, 97%);
-    border-block-start: 100px;
-    border-spacing: 10px;
-}
