--- src/components/Container.jsx
@@ -1,39 +1,0 @@
-import React from 'react'
-import "./Container.css"
-import starIcon from '../assets/images/icon-star.svg'
-import Faqs from './faq/Faqs'
-
-const faqs = [
-  {
-      question: "What is Frontend Mentor, and how will it help me?",
-      answer: "Frontend Mentor offers realistic coding challenges to help developers improve their frontend coding skills with projects in HTML, CSS, and JavaScript. It's suitable for all levels and ideal for portfolio building."
-  },
-  {   question: "Is Frontend Mentor free ?",
-      answer: "Yes, Frontend Mentor offers both free and premium coding challenges, with the free option providing access to a range of projects suitable for all skill levels."
-  },
-  {   question: "Can I use Frontend Mentor projects in my portfolio ?",
-      answer: "Yes, you can use projects completed on Frontend Mentor in your portfolio.It's an excellent way to showcase your skills to potential employers!"
-  },
-  {   question: "How can I get help if I'm stuck on a Frontend Mentor challenge?",
-      answer: "The best place to get help is inside Frontend Mentor's Discord community. There's a help channel where you can ask questions and seek support from other community members."
-  },
-]
-
-function Container() {
-  return (
-    <div className='container'>
-        <header>
-            <img src={starIcon} alt="Star Icon" />
-            <h1>FAQ</h1>
-        </header>
-        <ul className='faq-list'>
-          <Faqs {...faqs[0]}/>  
-          <Faqs {...faqs[1]}/>   
-          <Faqs {...faqs[2]}/>   
-          <Faqs {...faqs[3]}/>         
-        </ul>
-    </div>
-  )
-}
-
-export default Container
