+++ src/components/faq/Container.css
@@ -1,0 +1,82 @@
+.container {
+    max-width: 50%;
+    border-radius: 20px;
+    background-color: hsl(0, 0%, 100%);
+    box-shadow: rgba(100, 100, 111, 0.2) 0px 7px 29px 0px;
+    padding: 25px 50px;
+    height: auto;
+    max-height: 90%;
+}
+
+header {
+    font-size: 7rem;
+    display: inline-flex;
+    width: 100%;
+    align-items: center;
+}
+
+header img {
+    height: 4rem;
+    width: 2.5rem;
+    margin-right: 2.5rem;
+}
+
+header h1 {
+    font-size: 4rem;
+}
+
+.faq-list{
+    width: 100%;
+    height: 25%;
+    list-style: none;
+    display: grid;
+    grid-gap: 10px;
+    padding-left: 0px;
+    margin-bottom: 2rem;
+}
+
+.faq-list article:not(:last-child) {
+    border-bottom: 2px solid hsl(275, 100%, 97%);
+    border-block-start: 100px;
+    border-spacing: 10px;
+}
+
+@media screen and (max-width: 800px) {
+    .container {
+        padding:30px;
+        width: 100%;
+        max-width: 65%;
+        height: auto;
+    }
+
+    header {
+        font-size: 12px;
+    }
+}
+
+@media screen and (max-width: 600px) {
+    .container {
+        padding:25px;
+        width: 100%;
+        max-width: 80%;
+        height: auto;
+    }
+
+    header {
+        font-size: 2rem;
+    }
+
+    header h1{
+        font-size: 2.5rem;
+        margin:12px 0px;
+    }
+
+    header img {
+        height: 3rem;
+        width: 2.5rem;
+    }
+
+    .faq-list {
+        font-size:20px;
+    }
+}
