+++ src/components/faq/Faqs.css
@@ -1,0 +1,25 @@
+.faq-container {
+  width: 100vw;
+  height: 100vh;
+  background: url("../../assets/images/background-pattern-desktop.svg");
+  background-position: top;
+  background-size: 100%;
+  background-repeat: no-repeat;
+  background-color: hsl(275, 100%, 97%);
+}
+
+.grid {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+@media screen and (max-width: 600px) {
+    .faq-container {
+        background-image: url("../../assets/images/background-pattern-mobile.svg");
+        background-repeat: no-repeat;
+        background-position: top;
+        background-size: 100%;
+        height: 100vh;
+    }
+}
